Support private session keys for xoring with the queryid
This allows the export of semi-stable (but definitely un-stable) query
identifiers, without leaking information to clients (modulo their own
backend-linked C programs) to let them rely on long-term hash value
stability.

Also, save/restore under-estimation information after all

As long as a lot of mucking around is going on the file format, may as
well save/restore.

+++ contrib/pg_stat_statements/pg_stat_statements--1.2.sql
@@ -14,7 +14,7 @@ CREATE FUNCTION pg_stat_statements(
     OUT dbid oid,
     OUT stat_session_id int4,
     OUT query text,
-    OUT query_id int4,
+    OUT query_id int8,
     OUT calls int8,
     OUT calls_underest int8,
     OUT total_time float8,
